backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-Cors==4.0.0

# Data Handling and Numerics
pandas==2.2.1
numpy==1.26.4

# Machine Learning
scikit-learn==1.4.1.post1

# AI & Language Models (Google Gemini)
google-generativeai==0.4.1

# Graph Database Connector (Neo4j)
neo4j==5.17.0

# PDF Generation
fpdf2==2.7.8

# Synthetic Data Generation
Faker==24.4.0

# Environment Variable Management
python-dotenv==1.0.1